+++ Changes.md
@@ -4,3 +4,4 @@ Version 3.0.0.1
   1. Clone from CleanM theme: https://moodle.org/mod/forum/discuss.php?d=311119.
   2. TravisCI fixes.
   3. Fix no page URL on purge all caches.
+  4. Change renderer to core_renderer.